SFX
-For the coffee maker, Machine_Break_Pop plays when it breaks, and
Machine_Broken_Loop plays as long as it’s broken.
-Machine_Button_Press plays when the player presses the button to
dispense the coffee, and Machine_Coffee_Pour plays for as long as the
coffee is being dispensed. When the coffee is done being dispensed it
stops and Machine_Button_Release plays.
-Cup_Place plays when the cup is placed in the coffee maker, and
Cup_Remove when the cup is taken out of the coffee maker. Cup_Remove
may also play when the player trashes a cup.
-Machine_Voice_Broken plays as long as the coffee maker is broken.
-Machine_Voice_Leak plays as long as there is a leak.
-Machine_Voice_No_Beans plays when the coffee maker is out of beans.
-the Machine_Voice_Button sounds play whenever a button on the coffee
maker is pressed. Perhaps one is chosen randomly